requirements.txt
anthropic
openpyxl
pandas
Pillow
playwright
python-docx
python-dotenv
reportlab
